Add looping to fetch all the logs and improve debugging output
In summary the issue arises when the packer build output is very long.
The `logText/progressiveText` endpoint only gives us a few lines upto a
point so we have to loop through the output using the `start` query
parameter until we have what we need.

The `start` query parameter serves as an offset and the response has
`text-size` to mark the endpoint of the current output and `more-data`
to communicate if there's more data. If you [look at the code][1] it's
pretty easy to understand what's going on. This is not officially
documented but [someone online mentioned it as a bug][2] (I don't think
it is) that should have been fixed by now (seems incorrect since we have
to deal with it).

[1]: https://gist.github.com/carsongee/0ed0613eaa2d26eed437
[2]: https://stackoverflow.com/a/27108908/2184155